+++ source/define/constant/BOOLEAN.ts
@@ -1,0 +1,4 @@
+/**
+  It represents boolean values.
+*/
+export const BOOLEAN: string = 'boolean';