+++ source/define/constant/NUMBER.ts
@@ -1,0 +1,4 @@
+/**
+  It represents numerical values.
+*/
+export const NUMBER: string = 'number';