+++ source/define/constant/STRING.ts
@@ -1,0 +1,4 @@
+/**
+  It represents text values.
+*/
+export const STRING: string = 'string';