+++ source/define/constant/UNDEFINED.ts
@@ -1,0 +1,4 @@
+/**
+  It represents the undefined value.
+*/
+export const UNDEFINED: string = 'undefined';